Move applyAclToEpisode() to endpoint
...to avoid dependency loops between the AssetManager and the
AclServiceFactory in the future. This also keeps the AclService cleaner
since it now only deals with ACL templates.
