+++ modules/authorization-manager/src/main/java/org/opencastproject/authorization/xacml/manager/endpoint/AbstractAclServiceRestEndpoint.java
@@ -27,6 +27,7 @@ import static javax.servlet.http.HttpServletResponse.SC_INTERNAL_SERVER_ERROR;
 import static javax.servlet.http.HttpServletResponse.SC_NOT_FOUND;
 import static javax.servlet.http.HttpServletResponse.SC_OK;
 import static org.opencastproject.authorization.xacml.manager.util.Util.getManagedAcl;
+import static org.opencastproject.authorization.xacml.manager.util.Util.toAcl;
 import static org.opencastproject.util.RestUtil.R.conflict;
 import static org.opencastproject.util.RestUtil.R.noContent;
 import static org.opencastproject.util.RestUtil.R.notFound;
@@ -38,14 +39,19 @@ import static org.opencastproject.util.doc.rest.RestParameter.Type.BOOLEAN;
 import static org.opencastproject.util.doc.rest.RestParameter.Type.INTEGER;
 import static org.opencastproject.util.doc.rest.RestParameter.Type.STRING;
 
+import org.opencastproject.assetmanager.api.AssetManager;
 import org.opencastproject.authorization.xacml.manager.api.AclService;
 import org.opencastproject.authorization.xacml.manager.api.AclServiceException;
 import org.opencastproject.authorization.xacml.manager.api.AclServiceFactory;
 import org.opencastproject.authorization.xacml.manager.api.ManagedAcl;
 import org.opencastproject.authorization.xacml.manager.impl.ManagedAclImpl;
+import org.opencastproject.mediapackage.MediaPackage;
+import org.opencastproject.mediapackage.MediaPackageException;
 import org.opencastproject.security.api.AccessControlList;
 import org.opencastproject.security.api.AccessControlParser;
 import org.opencastproject.security.api.AccessControlUtil;
+import org.opencastproject.security.api.AclScope;
+import org.opencastproject.security.api.AuthorizationService;
 import org.opencastproject.security.api.Organization;
 import org.opencastproject.security.api.SecurityService;
 import org.opencastproject.util.Jsons;
@@ -57,6 +63,8 @@ import org.opencastproject.util.data.functions.Options;
 import org.opencastproject.util.doc.rest.RestParameter;
 import org.opencastproject.util.doc.rest.RestQuery;
 import org.opencastproject.util.doc.rest.RestResponse;
+
+import com.entwinemedia.fn.data.Opt;
 
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
@@ -82,6 +90,10 @@ public abstract class AbstractAclServiceRestEndpoint {
   protected abstract String getEndpointBaseUrl();
 
   protected abstract SecurityService getSecurityService();
+
+  protected abstract AssetManager getAssetManager();
+
+  protected abstract AuthorizationService getAuthorizationService();
 
   @GET
   @Path("/acl/{aclId}")
@@ -326,12 +338,38 @@ public abstract class AbstractAclServiceRestEndpoint {
       return notFound();
     }
     try {
-      if (aclService.applyAclToEpisode(episodeId, Options.join(macl))) {
+      Option<AccessControlList> aclOpt = Options.join(macl).map(toAcl);
+      Opt<MediaPackage> mediaPackage = getAssetManager().getMediaPackage(episodeId);
+      // the episode service is the source of authority for the retrieval of media packages
+      if (mediaPackage.isSome()) {
+        MediaPackage episodeSvcMp = mediaPackage.get();
+        aclOpt.fold(new Option.EMatch<AccessControlList>() {
+          // set the new episode ACL
+          @Override
+          public void esome(final AccessControlList acl) {
+            // update in episode service
+            try {
+              MediaPackage mp = getAuthorizationService().setAcl(episodeSvcMp, AclScope.Episode, acl).getA();
+              getAssetManager().takeSnapshot(mp);
+            } catch (MediaPackageException e) {
+              logger.error("Error getting ACL from media package", e);
+            }
+          }
+
+          // if none EpisodeACLTransition#isDelete returns true so delete the episode ACL
+          @Override
+          public void enone() {
+            // update in episode service
+            MediaPackage mp = getAuthorizationService().removeAcl(episodeSvcMp, AclScope.Episode);
+            getAssetManager().takeSnapshot(mp);
+          }
+
+        });
         return ok();
-      } else {
-        return notFound();
       }
-    } catch (AclServiceException e) {
+      // not found
+      return notFound();
+    } catch (Exception e) {
       logger.error("Error applying acl to episode {}", episodeId);
       return serverError();
     }
