+++ modules/authorization-manager/src/main/java/org/opencastproject/authorization/xacml/manager/impl/AclServiceImpl.java
@@ -21,9 +21,6 @@
 
 package org.opencastproject.authorization.xacml.manager.impl;
 
-import static org.opencastproject.authorization.xacml.manager.util.Util.toAcl;
-
-import org.opencastproject.assetmanager.api.AssetManager;
 import org.opencastproject.authorization.xacml.manager.api.AclService;
 import org.opencastproject.authorization.xacml.manager.api.AclServiceException;
 import org.opencastproject.authorization.xacml.manager.api.ManagedAcl;
@@ -35,18 +32,12 @@ import org.opencastproject.elasticsearch.index.event.Event;
 import org.opencastproject.elasticsearch.index.event.EventSearchQuery;
 import org.opencastproject.elasticsearch.index.series.Series;
 import org.opencastproject.elasticsearch.index.series.SeriesSearchQuery;
-import org.opencastproject.mediapackage.MediaPackage;
-import org.opencastproject.mediapackage.MediaPackageException;
 import org.opencastproject.security.api.AccessControlList;
-import org.opencastproject.security.api.AclScope;
-import org.opencastproject.security.api.AuthorizationService;
 import org.opencastproject.security.api.Organization;
 import org.opencastproject.security.api.SecurityService;
 import org.opencastproject.security.api.User;
 import org.opencastproject.util.NotFoundException;
 import org.opencastproject.util.data.Option;
-
-import com.entwinemedia.fn.data.Opt;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -65,76 +56,19 @@ public final class AclServiceImpl implements AclService {
 
   /** Service dependencies */
   private final AclDb aclDb;
-  private final AssetManager assetManager;
-  private final AuthorizationService authorizationService;
   private final SecurityService securityService;
 
   /** The Elasticsearch indices */
   protected AbstractSearchIndex adminUiIndex;
   protected AbstractSearchIndex externalApiIndex;
 
-  public AclServiceImpl(Organization organization, AclDb aclDb, AssetManager assetManager,
-          AuthorizationService authorizationService, AbstractSearchIndex adminUiIndex,
+  public AclServiceImpl(Organization organization, AclDb aclDb, AbstractSearchIndex adminUiIndex,
           AbstractSearchIndex externalApiIndex, SecurityService securityService) {
     this.organization = organization;
     this.aclDb = aclDb;
-    this.assetManager = assetManager;
-    this.authorizationService = authorizationService;
     this.adminUiIndex = adminUiIndex;
     this.externalApiIndex = externalApiIndex;
     this.securityService = securityService;
-  }
-
-  @Override
-  public boolean applyAclToEpisode(String episodeId, AccessControlList acl) throws AclServiceException {
-    try {
-      Opt<MediaPackage> mediaPackage = Opt.none();
-      if (assetManager != null) {
-        mediaPackage = assetManager.getMediaPackage(episodeId);
-      }
-
-      Option<AccessControlList> aclOpt = Option.option(acl);
-      // the episode service is the source of authority for the retrieval of media packages
-      if (mediaPackage.isSome()) {
-        MediaPackage episodeSvcMp = mediaPackage.get();
-        aclOpt.fold(new Option.EMatch<AccessControlList>() {
-          // set the new episode ACL
-          @Override
-          public void esome(final AccessControlList acl) {
-            // update in episode service
-            try {
-              MediaPackage mp = authorizationService.setAcl(episodeSvcMp, AclScope.Episode, acl).getA();
-              if (assetManager != null) {
-                assetManager.takeSnapshot(mp);
-              }
-            } catch (MediaPackageException e) {
-              logger.error("Error getting ACL from media package", e);
-            }
-          }
-
-          // if none EpisodeACLTransition#isDelete returns true so delete the episode ACL
-          @Override
-          public void enone() {
-            // update in episode service
-            MediaPackage mp = authorizationService.removeAcl(episodeSvcMp, AclScope.Episode);
-            if (assetManager != null) {
-              assetManager.takeSnapshot(mp);
-            }
-          }
-
-        });
-        return true;
-      }
-      // not found
-      return false;
-    } catch (Exception e) {
-      throw new AclServiceException(e);
-    }
-  }
-
-  @Override
-  public boolean applyAclToEpisode(String episodeId, Option<ManagedAcl> managedAcl) throws AclServiceException {
-    return applyAclToEpisode(episodeId, managedAcl.map(toAcl).getOrElseNull());
   }
 
   @Override
