+++ modules/authorization-manager/src/main/resources/OSGI-INF/acl-service-rest.xml
@@ -12,11 +12,11 @@
   </service>
   <reference name="acl-service-factory"
              interface="org.opencastproject.authorization.xacml.manager.api.AclServiceFactory"
-             cardinality="1..1" policy="static" bind="setAclServiceFactory"/>
+             bind="setAclServiceFactory"/>
   <reference name="security-service" interface="org.opencastproject.security.api.SecurityService"
-             cardinality="1..1" policy="static" bind="setSecurityService"/>
+             bind="setSecurityService"/>
   <reference name="authorization-service" interface="org.opencastproject.security.api.AuthorizationService"
-             cardinality="1..1" policy="static" bind="setAuthorizationService"/>
+             bind="setAuthorizationService"/>
   <reference name="asset-manager" interface="org.opencastproject.assetmanager.api.AssetManager"
-             cardinality="1..1" policy="dynamic" bind="setAssetManager"/>
+             bind="setAssetManager"/>
 </scr:component>
