+++ modules/authorization-manager/src/test/java/org/opencastproject/authorization/xacml/manager/impl/AclScannerTest.java
@@ -105,7 +105,7 @@ public class AclScannerTest {
     AclServiceFactory aclServiceFactory = new AclServiceFactory() {
       @Override
       public AclService serviceFor(Organization org) {
-        return new AclServiceImpl(new DefaultOrganization(), aclDb, null, null, index, index, securityService);
+        return new AclServiceImpl(new DefaultOrganization(), aclDb, index, index, securityService);
       }
     };
